refactor(Peer Group): Move PeerGroupWork api to new class
